Always use color
I personally hold the opinion that tools should output as much as possible
and output can be piped through another stage that strips color if color
stripping is desired - why is it the compilers job to work out whether or
not it's in a terminal?

I also understand that convention has settled against this, and so would
not oppose adding terminal detection to the scons build system if we must.