Fix up local pipes and activate them
These may prove to be a pain in the arse in future, but for now it
may be worth keeping these around. In future if we ever come up
with a sane single-player mode (har har... but a tutorial could be
nice here), it would be nice to not set off any firewall warnings.

But the primary focus is multiplayer so firewall warnings are a bit
of a moot point.

Either way, it works, and it's not a mess. I think.