Preliminary proof that Windows is an awful OS
That is, preliminary implementation of the networking stuff.

The Win32 stuff is untested as I haven't installed the necessary
headers and libs for SDL2 for MinGW. Everything compiles up to but
not including linking it all together.

As a side effect, Win32 lacks the ability to pass `MSG_DONTWAIT` to
`send()` or `recv()`, so I have to make the sockets non-blocking.
Perhaps it wasn't such a great idea to emulate blocking for
`connect()`, but at the same time, I now know that it's possible to
do a non-blocking `connect()`, at least on Linux.

Oh yeah, also, thanks to Windows not being designed for networking,
Winsock provides its own error handling separate from `errno`, so
you can guarantee that it's actually going to crash due to it not
specially handling the Super Special(tm) special snowflake
specialness special error codes required by Winsock.

More special cases, such as things like `SD_BOTH` instead of
`SHUT_RDWR` when calling `shutdown()`, `WSAPoll()` instead of
`poll()` (which I *think* was added in Vista?), the classic
`closesocket()` and `ioctlsocket()` theatre that you have to deal
with.

Did I mention `WSAStartup()`? Because you need to call that.
Apparently you also need to call `WSACleanup()`. I'd be tempted to
avoid having to do that, but I at least want this to be *somewhat*
respectful to people who use Windows.
(In other words, this isn't Firefox.)

The `TODO: Abolish Windows` comment stands.