Remove exceptions from the code because Windows
The Windows builds I've been doing for testing have had this
*slight* problem that any exception you catch never actually gets
caught. It bubbles up all the way

We're going to need a Windows maintainer at this rate.

Either that, or the Windows build will be a VirtualBox or QEMU VM
with a lightweight Linux distro and the game preinstalled.

I'd rather handle all the BSDs I know of than deal with Windows.
That is: FreeBSD, NetBSD, OpenBSD, DragonflyBSD, and MirBSD,
and the latter two I have never touched in my life.

Or, y'know, we could port this to DOS and add support for DOSBox's
IPX over UDP/IP tunnels. The protocol's pretty easy to deal with,
and the DOS side isn't bad either. We'd definitely need to support
transport over UDP, but once that's sorted, converting that to IPX
should be pretty easy.

I suspect the way forward is going to be the Linux VM approach.
Bundle QEMU with it and you've got a Windows release.

It'll still be smaller than your average Denuvo-laden PoS anyway.